backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
psycopg2-binary
pyjwt